requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cmake==3.18.4.post1
cycler==0.10.0
distro==1.5.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.3
pytz==2021.1
scikit-build==0.11.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tornado==6.1
tzlocal==2.1
